requirements.txt
streamlit 
pandas 
tabula-py 
openpyxl
